Replace Update_Book_Copy_Status process with BorrowLogEventTrigger
Activated change data capture on the Borrow Log object and created a
change event trigger on Borrow_Log_ChangeEvent. When a Borrow Log change
event is published, the BorrowLogEvenTrigger is executed asynchronously.
The trigger updates associated Book Copy statuses to 'Loaned' when the
change event type is 'CREATE', and 'Returned' when the event type is
'UPDATE'. A test class is included that asserts that the copy status is
correctly updated when the corresponding change events occur on Borrow
Log.

Pending
 *  Remove logic from the BorrowLogEventTrigger